Move more jobs functions out of views.py
