+++ core/views/views.py
@@ -2,17 +2,15 @@
 from __future__ import unicode_literals
 
 # generic
-import ast
 import json
 import logging
 import re
-from urllib.parse import urlencode
 
 # django
 from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
 from django.db.models.query import QuerySet
-from django.http import HttpResponse, JsonResponse, FileResponse
+from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render, redirect
 
 # import models
@@ -20,9 +18,6 @@ from core import models
 
 # import oai server
 from core.oai import OAIProvider
-
-# import background tasks
-from core import tasks
 
 # import mongo dependencies
 from core.mongo import *
@@ -403,422 +398,6 @@ def system_bg_status(request):
     })
 
 
-####################################################################
-# Job Validation Report       									   #
-####################################################################
-
-@login_required
-def job_reports_create_validation(request, org_id, record_group_id, job_id):
-    """
-    Generate job report based on validation results
-    """
-
-    # retrieve job
-    cjob = models.CombineJob.get_combine_job(int(job_id))
-
-    # if GET, prepare form
-    if request.method == 'GET':
-
-        # mapped field analysis, generate if not part of job_details
-        if 'mapped_field_analysis' in cjob.job.job_details_dict.keys():
-            field_counts = cjob.job.job_details_dict['mapped_field_analysis']
-        else:
-            if cjob.job.finished:
-                field_counts = cjob.count_indexed_fields()
-                cjob.job.update_job_details({'mapped_field_analysis': field_counts}, save=True)
-            else:
-                logger.debug('job not finished, not setting')
-                field_counts = {}
-
-        # render page
-        return render(request, 'core/job_reports_create_validation.html', {
-            'cjob': cjob,
-            'field_counts': field_counts,
-            'breadcrumbs': breadcrumb_parser(request)
-        })
-
-    # if POST, generate report
-    if request.method == 'POST':
-
-        # get job name for Combine Task
-        report_name = request.POST.get('report_name')
-        if report_name == '':
-            report_name = 'j_%s_validation_report' % cjob.job.id
-            combine_task_name = "Validation Report: %s" % cjob.job.name
-        else:
-            combine_task_name = "Validation Report: %s" % report_name
-
-        # handle POST params and save as Combine task params
-        task_params = {
-            'job_id': cjob.job.id,
-            'report_name': report_name,
-            'report_format': request.POST.get('report_format'),
-            'compression_type': request.POST.get('compression_type'),
-            'validation_scenarios': request.POST.getlist('validation_scenario', []),
-            'mapped_field_include': request.POST.getlist('mapped_field_include', [])
-        }
-
-        # cast to int
-        task_params['validation_scenarios'] = [int(vs_id) for vs_id in task_params['validation_scenarios']]
-
-        # remove select, reserved fields if in mapped field request
-        task_params['mapped_field_include'] = [f for f in task_params['mapped_field_include'] if
-                                               f not in ['record_id', 'db_id', 'oid', '_id']]
-
-        # initiate Combine BG Task
-        ct = models.CombineBackgroundTask(
-            name=combine_task_name,
-            task_type='validation_report',
-            task_params_json=json.dumps(task_params)
-        )
-        ct.save()
-
-        # run celery task
-        background_task = tasks.create_validation_report.delay(ct.id)
-        logger.debug('firing bg task: %s' % background_task)
-        ct.celery_task_id = background_task.task_id
-        ct.save()
-
-        # redirect to Background Tasks
-        return redirect('bg_tasks')
-
-
-@login_required
-def job_update(request, org_id, record_group_id, job_id):
-    """
-    Update Job in one of several ways:
-        - re-map and index
-        - run new / different validations
-    """
-
-    # retrieve job
-    cjob = models.CombineJob.get_combine_job(int(job_id))
-
-    # if GET, prepare form
-    if request.method == 'GET':
-        # get validation scenarios
-        validation_scenarios = models.ValidationScenario.objects.all()
-
-        # get field mappers
-        field_mappers = models.FieldMapper.objects.all()
-        orig_fm_config_json = cjob.job.get_fm_config_json()
-
-        # get all bulk downloads
-        bulk_downloads = models.DPLABulkDataDownload.objects.all()
-
-        # get update type from GET params
-        update_type = request.GET.get('update_type', None)
-
-        # render page
-        return render(request, 'core/job_update.html', {
-            'cjob': cjob,
-            'update_type': update_type,
-            'validation_scenarios': validation_scenarios,
-            'field_mappers': field_mappers,
-            'bulk_downloads': bulk_downloads,
-            'xml2kvp_handle': models.XML2kvp(),
-            'orig_fm_config_json': orig_fm_config_json,
-            'breadcrumbs': breadcrumb_parser(request)
-        })
-
-    # if POST, submit job
-    if request.method == 'POST':
-
-        logger.debug('updating job')
-        logger.debug(request.POST)
-
-        # retrieve job
-        cjob = models.CombineJob.get_combine_job(int(job_id))
-
-        # get update type
-        update_type = request.POST.get('update_type', None)
-        logger.debug('running job update: %s' % update_type)
-
-        # handle re-index
-        if update_type == 'reindex':
-            # get preferred metadata index mapper
-            fm_config_json = request.POST.get('fm_config_json')
-
-            # init re-index
-            cjob.reindex_bg_task(fm_config_json=fm_config_json)
-
-            # set gms
-            gmc = models.GlobalMessageClient(request.session)
-            gmc.add_gm({
-                'html': '<p><strong>Re-Indexing Job:</strong><br>%s</p>'
-                        '<p><a href="%s"><button type="button" '
-                        'class="btn btn-outline-primary btn-sm">View Background Tasks</button></a></p>' % (
-                            cjob.job.name, reverse('bg_tasks')),
-                'class': 'success'
-            })
-
-            return redirect('job_details',
-                            org_id=cjob.job.record_group.organization.id,
-                            record_group_id=cjob.job.record_group.id,
-                            job_id=cjob.job.id)
-
-        # handle new validations
-        if update_type == 'validations':
-            # get requested validation scenarios
-            validation_scenarios = request.POST.getlist('validation_scenario', [])
-
-            # get validations
-            validations = models.ValidationScenario.objects.filter(
-                id__in=[int(vs_id) for vs_id in validation_scenarios])
-
-            # init bg task
-            cjob.new_validations_bg_task([vs.id for vs in validations])
-
-            # set gms
-            gmc = models.GlobalMessageClient(request.session)
-            gmc.add_gm({
-                'html': '<p><strong>Running New Validations for Job:</strong><br>%s<br>'
-                        '<br><strong>Validation Scenarios:</strong><br>%s</p>'
-                        '<p><a href="%s"><button type="button" '
-                        'class="btn btn-outline-primary btn-sm">View Background Tasks</button></a></p>' % (
-                            cjob.job.name, '<br>'.join([vs.name for vs in validations]), reverse('bg_tasks')),
-                'class': 'success'
-            })
-
-            return redirect('job_details',
-                            org_id=cjob.job.record_group.organization.id,
-                            record_group_id=cjob.job.record_group.id,
-                            job_id=cjob.job.id)
-
-        # handle validation removal
-        if update_type == 'remove_validation':
-            # get validation scenario to remove
-            jv_id = request.POST.get('jv_id', False)
-
-            # initiate Combine BG Task
-            cjob.remove_validation_bg_task(jv_id)
-
-            # set gms
-            vs = models.JobValidation.objects.get(pk=int(jv_id)).validation_scenario
-            gmc = models.GlobalMessageClient(request.session)
-            gmc.add_gm({
-                'html': '<p><strong>Removing Validation for Job:</strong><br>%s<br><br>'
-                        '<strong>Validation Scenario:</strong><br>%s</p><p><a href="%s"><button type="button" '
-                        'class="btn btn-outline-primary btn-sm">View Background Tasks</button></a></p>' % (
-                            cjob.job.name, vs.name, reverse('bg_tasks')),
-                'class': 'success'
-            })
-
-            return redirect('job_details',
-                            org_id=cjob.job.record_group.organization.id,
-                            record_group_id=cjob.job.record_group.id,
-                            job_id=cjob.job.id)
-
-        # handle validation removal
-        if update_type == 'dbdm':
-            # get validation scenario to remove
-            dbdd_id = request.POST.get('dbdd', False)
-
-            # initiate Combine BG Task
-            cjob.dbdm_bg_task(dbdd_id)
-
-            # set gms
-            dbdd = models.DPLABulkDataDownload.objects.get(pk=int(dbdd_id))
-            gmc = models.GlobalMessageClient(request.session)
-            gmc.add_gm({
-                'html': '<p><strong>Running DPLA Bulk Data comparison for Job:</strong><br>%s<br><br>'
-                        '<strong>Bulk Data S3 key:</strong><br>%s</p><p><a href="%s"><button type="button" '
-                        'class="btn btn-outline-primary btn-sm">View Background Tasks</button></a></p>' % (
-                            cjob.job.name, dbdd.s3_key, reverse('bg_tasks')),
-                'class': 'success'
-            })
-
-            return redirect('job_details',
-                            org_id=cjob.job.record_group.organization.id,
-                            record_group_id=cjob.job.record_group.id,
-                            job_id=cjob.job.id)
-
-
-####################################################################
-# Job Validation Report       									   #
-####################################################################
-
-def document_download(request):
-    """
-    Args (GET params):
-        file_location: location on disk for file
-        file_download_name: desired download name
-        content_type: ContentType Headers
-    """
-
-    # known download format params
-    download_format_hash = {
-        'excel': {
-            'extension': '.xlsx',
-            'content_type': 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
-        },
-        'csv': {
-            'extension': '.csv',
-            'content_type': 'text/plain'
-        },
-        'tsv': {
-            'extension': '.tsv',
-            'content_type': 'text/plain'
-        },
-        'json': {
-            'extension': '.json',
-            'content_type': 'text/plain'
-        },
-        'zip': {
-            'extension': '.zip',
-            'content_type': 'application/zip'
-        },
-        'targz': {
-            'extension': '.tar.gz',
-            'content_type': 'application/gzip'
-        }
-    }
-
-    # get params
-    download_format = request.GET.get('download_format', None)
-    filepath = request.GET.get('filepath', None)
-    name = request.GET.get('name', 'download')
-    # TODO: content_type is never used
-    content_type = request.GET.get('content_type', 'text/plain')
-    preview = request.GET.get('preview', False)
-
-    # if known download format, use hash and overwrite provided or defaults
-    if download_format and download_format in download_format_hash.keys():
-        format_params = download_format_hash[download_format]
-        name = '%s%s' % (name, format_params['extension'])
-        content_type = format_params['content_type']
-
-    # generate response
-    response = FileResponse(open(filepath, 'rb'))
-    if not preview:
-        response['Content-Disposition'] = 'attachment; filename="%s"' % name
-    return response
-
-
-####################################################################
-# Jobs QA	                   									   #
-####################################################################
-
-@login_required
-def field_analysis(request, es_index):
-    # get field name
-    field_name = request.GET.get('field_name')
-
-    # get ESIndex, evaluating stringified list
-    esi = models.ESIndex(ast.literal_eval(es_index))
-
-    # get analysis for field
-    field_metrics = esi.field_analysis(field_name, metrics_only=True)
-
-    # return
-    return render(request, 'core/field_analysis.html', {
-        'esi': esi,
-        'field_name': field_name,
-        'field_metrics': field_metrics,
-        'breadcrumbs': breadcrumb_parser(request)
-    })
-
-
-@login_required
-def job_indexing_failures(request, org_id, record_group_id, job_id):
-    # get CombineJob
-    cjob = models.CombineJob.get_combine_job(job_id)
-
-    # return
-    return render(request, 'core/job_indexing_failures.html', {
-        'cjob': cjob,
-        'breadcrumbs': breadcrumb_parser(request)
-    })
-
-
-@login_required
-def field_analysis_docs(request, es_index, filter_type):
-    """
-
-    Table of documents that match a filtered ES query.
-
-    Args:
-        es_index (str): string ES index name
-        filter_type (str): what kind of filtering to impose on documents returned
-    """
-
-    # regardless of filtering type, get field name
-    field_name = request.GET.get('field_name')
-
-    # get ESIndex
-    esi = models.ESIndex(ast.literal_eval(es_index))
-
-    # begin construction of DT GET params with 'fields_names'
-    dt_get_params = [
-        ('field_names', 'db_id'),  # get DB ID
-        ('field_names', 'combine_id'),  # get Combine ID
-        ('field_names', 'record_id'),  # get ID from ES index document
-        ('field_names', field_name),  # add field to returned fields
-        ('filter_field', field_name),
-        ('filter_type', filter_type)
-    ]
-
-    # analysis scenario dict
-    analysis_scenario = {
-        'exists': None,
-        'matches': None,
-        'value': None
-    }
-
-    # field existence
-    if filter_type == 'exists':
-        # if check exists, get expected GET params
-        exists = request.GET.get('exists')
-        dt_get_params.append(('exists', exists))
-
-        # update analysis scenario dict
-        analysis_scenario['exists'] = exists
-
-    # field equals
-    if filter_type == 'equals':
-
-        # if check equals, get expected GET params
-        matches = request.GET.get('matches')
-        dt_get_params.append(('matches', matches))
-
-        value = request.GET.get('value', None)  # default None if checking non-matches to value
-        if value:
-            dt_get_params.append(('filter_value', value))
-
-        # update analysis scenario dict
-        analysis_scenario['matches'] = matches
-        analysis_scenario['value'] = value
-
-    # construct DT Ajax GET parameters string from tuples
-    dt_get_params_string = urlencode(dt_get_params)
-
-    # return
-    return render(request, 'core/field_analysis_docs.html', {
-        'esi': esi,
-        'field_name': field_name,
-        'filter_type': filter_type,
-        'analysis_scenario': analysis_scenario,
-        'msg': None,
-        'dt_get_params_string': dt_get_params_string,
-        'breadcrumbs': breadcrumb_parser(request)
-    })
-
-
-@login_required
-def job_validation_scenario_failures(request, org_id, record_group_id, job_id, job_validation_id):
-    # get CombineJob
-    cjob = models.CombineJob.get_combine_job(job_id)
-
-    # get job validation instance
-    jv = models.JobValidation.objects.get(pk=int(job_validation_id))
-
-    # return
-    return render(request, 'core/job_validation_scenario_failures.html', {
-        'cjob': cjob,
-        'jv': jv,
-        'breadcrumbs': breadcrumb_parser(request)
-    })
-
 
 ####################################################################
 # Configuration 												   #
@@ -959,115 +538,6 @@ def search(request):
     })
 
 
-####################################################################
-# Analysis  													   #
-####################################################################
-
-def analysis(request):
-    """
-    Analysis home
-    """
-
-    # get all jobs associated with record group
-    analysis_jobs = models.Job.objects.filter(job_type='AnalysisJob')
-
-    # get analysis jobs hierarchy
-    analysis_hierarchy = models.AnalysisJob.get_analysis_hierarchy()
-
-    # get analysis jobs lineage
-    analysis_job_lineage = models.Job.get_all_jobs_lineage(
-        organization=analysis_hierarchy['organization'],
-        record_group=analysis_hierarchy['record_group'],
-        exclude_analysis_jobs=False
-    )
-
-    # loop through jobs
-    for job in analysis_jobs:
-        # update status
-        job.update_status()
-
-    # render page
-    return render(request, 'core/analysis.html', {
-        'jobs': analysis_jobs,
-        'job_lineage_json': json.dumps(analysis_job_lineage),
-        'for_analysis': True,
-        'breadcrumbs': breadcrumb_parser(request)
-    })
-
-
-@login_required
-def job_analysis(request):
-    """
-    Run new analysis job
-    """
-
-    # if GET, prepare form
-    if request.method == 'GET':
-
-        # retrieve jobs (limiting if needed)
-        input_jobs = models.Job.objects.all()
-
-        # limit if analysis_type set
-        analysis_type = request.GET.get('type', None)
-        subset = request.GET.get('subset', None)
-        if analysis_type == 'published':
-
-            # load PublishedRecords
-            published = models.PublishedRecords(subset=subset)
-
-            # define input_jobs
-            input_jobs = published.published_jobs
-
-        else:
-            published = None
-
-        # get validation scenarios
-        validation_scenarios = models.ValidationScenario.objects.all()
-
-        # get field mappers
-        field_mappers = models.FieldMapper.objects.all()
-
-        # get record identifier transformation scenarios
-        rits = models.RecordIdentifierTransformationScenario.objects.all()
-
-        # get job lineage for all jobs (filtered to input jobs scope)
-        ld = models.Job.get_all_jobs_lineage(jobs_query_set=input_jobs)
-
-        # get all bulk downloads
-        bulk_downloads = models.DPLABulkDataDownload.objects.all()
-
-        # render page
-        return render(request, 'core/job_analysis.html', {
-            'job_select_type': 'multiple',
-            'input_jobs': input_jobs,
-            'published': published,
-            'validation_scenarios': validation_scenarios,
-            'rits': rits,
-            'field_mappers': field_mappers,
-            'xml2kvp_handle': models.XML2kvp(),
-            'analysis_type': analysis_type,
-            'bulk_downloads': bulk_downloads,
-            'job_lineage_json': json.dumps(ld)
-        })
-
-    # if POST, submit job
-    if request.method == 'POST':
-
-        cjob = models.CombineJob.init_combine_job(
-            user=request.user,
-            record_group=record_group,
-            job_type_class=models.AnalysisJob,
-            job_params=request.POST)
-
-        # start job and update status
-        job_status = cjob.start_job()
-
-        # if job_status is absent, report job status as failed
-        if job_status is False:
-            cjob.job.status = 'failed'
-            cjob.job.save()
-
-        return redirect('analysis')
 
 
 ####################################################################
